# docs/api/openapi.yaml
openapi: 3.1.0
info:
  title: Hermes Mobile Control Plane Gateway API
  version: 0.1.0
  description: >
    Implementation-ready API contract for Hermes Control Gateway. The API is
    self-hosted first and intended to run beside each Hermes install. WebSocket
    event streaming is documented as an HTTP upgrade endpoint with event schemas.
servers:
  - url: https://{gateway_host}/v1
    description: Direct gateway over Tailscale, trusted local network, HTTPS, or future relay
    variables:
      gateway_host:
        default: 100.64.0.10:8787
security:
  - bearerAuth: []
...
paths:
...
  /pairing/sessions:
    post:
      tags: [pairing]
      summary: Create a short-lived pairing session
      description: Started from a trusted local admin context on the Hermes host.
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreatePairingSessionRequest"
      responses:
        "201":
          description: Pairing challenge created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PairingSession"
        "403":
          $ref: "#/components/responses/Forbidden"
  /pairing/start:
    post:
      tags: [pairing]
      summary: Create a short-lived pairing session
      description: Alias for `/pairing/sessions` used by the first executable slice.
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreatePairingSessionRequest"
      responses:
        "201":
          description: Pairing challenge created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PairingSession"
        "403":
          $ref: "#/components/responses/Forbidden"
  /pairing/sessions/{pairing_id}:
    get:
      tags: [pairing]
      summary: Read pairing session status
      security: []
      parameters:
        - $ref: "#/components/parameters/PairingId"
      responses:
        "200":
          description: Pairing status
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PairingSession"
        "404":
          $ref: "#/components/responses/NotFound"
  /pairing/complete:
    post:
      tags: [pairing]
      summary: Complete device pairing
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CompletePairingRequest"
      responses:
        "200":
          description: Device registered and node credentials issued
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CompletePairingResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
        "409":
          $ref: "#/components/responses/Conflict"
  /auth/token/refresh:
    post:
      tags: [auth]
      summary: Refresh a device-bound session token
      security: []
      parameters:
        - $ref: "#/components/parameters/DeviceIdHeader"
        - $ref: "#/components/parameters/DeviceSignatureHeader"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RefreshTokenRequest"
      responses:
        "200":
          description: New session tokens
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthTokenSet"
        "401":
          $ref: "#/components/responses/Unauthorized"
...
  /devices/{device_id}:
    delete:
      tags: [devices]
      summary: Revoke a device
      parameters:
        - $ref: "#/components/parameters/DeviceId"
        - $ref: "#/components/parameters/DeviceSignatureHeader"
      responses:
        "204":
          description: Device revoked
        "404":
          $ref: "#/components/responses/NotFound"
...
  /nodes/register:
    post:
      tags: [inventory]
      summary: Register or update the local Hermes node
      description: Hermes-facing skeleton endpoint for local node metadata registration.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/NodeRegistration"
      responses:
        "201":
          description: Node registered
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Node"
  /inventory:
    get:
      tags: [inventory]
      summary: Get node and agent inventory
      parameters:
        - name: include_stale
          in: query
          schema:
            type: boolean
            default: true
      responses:
        "200":
          description: Inventory snapshot
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Inventory"
  /nodes/{node_id}:
    get:
      tags: [inventory]
      summary: Get node detail
      parameters:
        - $ref: "#/components/parameters/NodeId"
      responses:
        "200":
          description: Node detail
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Node"
        "404":
          $ref: "#/components/responses/NotFound"
  /agents:
    get:
      tags: [inventory]
      summary: List agents
      parameters:
        - name: node_id
          in: query
          schema:
            type: string
        - name: status
          in: query
          schema:
            $ref: "#/components/schemas/AgentStatus"
      responses:
        "200":
          description: Agents
          content:
            application/json:
              schema:
                type: object
                required: [agents]
                properties:
                  agents:
                    type: array
                    items:
                      $ref: "#/components/schemas/Agent"
...
  /sessions:
    get:
      tags: [chat]
      summary: List sessions
      parameters:
        - name: node_id
          in: query
          schema:
            type: string
        - name: agent_id
          in: query
          schema:
            type: string
        - name: status
          in: query
          schema:
            $ref: "#/components/schemas/SessionStatus"
      responses:
        "200":
          description: Sessions
          content:
            application/json:
              schema:
                type: object
                required: [sessions]
                properties:
                  sessions:
                    type: array
                    items:
                      $ref: "#/components/schemas/Session"
...
  /approvals:
    get:
      tags: [approvals]
      summary: List approval requests
      parameters:
        - name: state
          in: query
          schema:
            $ref: "#/components/schemas/ApprovalState"
        - name: node_id
          in: query
          schema:
            type: string
        - name: agent_id
          in: query
          schema:
            type: string
      responses:
        "200":
          description: Approval requests
          content:
            application/json:
              schema:
                type: object
                required: [approvals]
                properties:
                  approvals:
                    type: array
                    items:
                      $ref: "#/components/schemas/ApprovalRequest"
  /approvals/{approval_id}:
    get:
      tags: [approvals]
      summary: Get approval detail
      parameters:
        - $ref: "#/components/parameters/ApprovalId"
      responses:
        "200":
          description: Approval detail
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApprovalRequest"
  /approvals/{approval_id}/decisions:
    post:
      tags: [approvals]
      summary: Submit a signed approval decision
      parameters:
        - $ref: "#/components/parameters/ApprovalId"
        - $ref: "#/components/parameters/DeviceIdHeader"
        - $ref: "#/components/parameters/DeviceSignatureHeader"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ApprovalDecisionRequest"
      responses:
        "200":
          description: Decision accepted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApprovalDecisionResponse"
        "409":
          $ref: "#/components/responses/Conflict"
...
  /notifications/mobile_notify:
    post:
      tags: [notifications]
      summary: Hermes-facing request to send a mobile notification
      description: Called by Hermes or the gateway adapter, not directly by arbitrary mobile clients.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/MobileNotifyRequest"
      responses:
        "202":
          description: Notification accepted or queued
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Notification"
        "422":
          $ref: "#/components/responses/Unprocessable"
  /notifications:
    get:
      tags: [notifications]
      summary: List notification records
      parameters:
        - name: category
          in: query
          schema:
            $ref: "#/components/schemas/NotificationCategory"
        - name: after
          in: query
          schema:
            type: string
      responses:
        "200":
          description: Notifications
          content:
            application/json:
              schema:
                type: object
                required: [notifications]
                properties:
                  notifications:
                    type: array
                    items:
                      $ref: "#/components/schemas/Notification"
  /audit/events:
    get:
      tags: [audit]
      summary: Query audit events
      description: Return append-only audit events for forensic review and mobile history views.
      parameters:
        - name: node_id
          in: query
          schema:
            type: string
        - name: agent_id
          in: query
          schema:
            type: string
        - name: session_id
          in: query
          schema:
            type: string
        - name: event_type
          in: query
          schema:
            type: string
        - name: after
          in: query
          schema:
            type: string
            description: Cursor or audit event ID to resume after
        - name: limit
          in: query
          schema:
            type: integer
            default: 100
            maximum: 1000
      responses:
        "200":
          description: Audit event page
          content:
            application/json:
              schema:
                type: object
                required: [audit_events]
                properties:
                  audit_events:
                    type: array
                    items:
                      $ref: "#/components/schemas/AuditEvent"
                  next_cursor:
                    type: string
  /events:
    get:
      tags: [events]
      summary: Backfill events by cursor
      parameters:
        - name: after
          in: query
          schema:
            type: string
        - name: limit
          in: query
          schema:
            type: integer
            default: 500
            maximum: 2000
      responses:
        "200":
          description: Event backfill
          content:
            application/json:
              schema:
                type: object
                required: [events]
                properties:
                  events:
                    type: array
                    items:
                      $ref: "#/components/schemas/EventEnvelope"
                  next_cursor:
                    type: string
  /events/stream:
    get:
      tags: [events]
      summary: WebSocket event stream
      description: >
        Upgrade to WebSocket. Query parameters may include `after`, `node_id`,
        `agent_id`, `session_id`, and repeated `event_type`. Server emits
        EventEnvelope messages.
      parameters:
        - name: after
          in: query
          schema:
            type: string
      responses:
        "101":
          description: Switching Protocols to WebSocket
        "401":
          $ref: "#/components/responses/Unauthorized"
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
...
    DeviceIdHeader:
      name: X-Hermes-Device-Id
      in: header
      required: true
      schema: {type: string}
...
    PairingId:
      name: pairing_id
      in: path
      required: true
      schema: {type: string}
...
  responses:
...
    Forbidden:
      description: Not authorized
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
...
  schemas:
...
    CreatePairingSessionRequest:
      type: object
      required: [display_name]
      properties:
        display_name:
          type: string
        requested_permissions:
          type: array
          items:
            $ref: "#/components/schemas/Permission"
        ttl_seconds:
          type: integer
          minimum: -1
          maximum: 3600
          description: Test/development override for pairing token expiry.
    PairingSession:
      type: object
      required: [pairing_id, challenge, status, node_id, node_fingerprint, expires_at]
      properties:
        pairing_id:
          type: string
        pairing_token:
          type: string
          description: One-time token returned only when creating a pairing session.
        challenge:
          type: string
        status:
          enum: [pending, completed, expired, cancelled]
        node_id:
          type: string
        node_fingerprint:
          type: string
        expires_at:
          type: string
          format: date-time
    CompletePairingRequest:
      type: object
      required: [pairing_id, challenge_response, device_public_key, device]
      properties:
        pairing_id:
          type: string
        challenge_response:
          type: string
        device_public_key:
          type: string
        device:
          $ref: "#/components/schemas/DeviceRegistration"
...